[feature] PHPUnit added and user unitary test suite added
