+++ buildspec.yml
@@ -7,4 +7,5 @@ phases:
     commands:
       - composer format
       - composer lint
-      - composer lint:fix    
+      - composer lint:fix  
+      - composer test
