+++ src/Presentation/Factories/AuthenticationControllerFactory.php
@@ -6,8 +6,8 @@ use Mvreisg\GamebaseBackend\Application\Services\AuthenticationService;
 use Mvreisg\GamebaseBackend\Infrastructure\Connections\MariaDBConnection;
 use Mvreisg\GamebaseBackend\Infrastructure\Connections\RedisConnection;
 use Mvreisg\GamebaseBackend\Infrastructure\Encryption\SodiumEncryption;
-use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDBUserRepository;
-use Mvreisg\GamebaseBackend\Infrastructure\Cache\RedisUserCache;
+use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDB\MariaDBUserRepository;
+use Mvreisg\GamebaseBackend\Infrastructure\Cache\Redis\RedisUserCache;
 use Mvreisg\GamebaseBackend\Presentation\Controllers\AuthenticationController;
 
 class AuthenticationControllerFactory
