+++ src/Presentation/Factories/GameGenreControllerFactory.php
@@ -4,13 +4,13 @@ namespace Mvreisg\GamebaseBackend\Presentation\Factories;
 
 use Mvreisg\GamebaseBackend\Application\Services\AuthenticationService;
 use Mvreisg\GamebaseBackend\Application\Services\GameGenreService;
-use Mvreisg\GamebaseBackend\Infrastructure\Cache\RedisUserCache;
+use Mvreisg\GamebaseBackend\Infrastructure\Cache\Redis\RedisUserCache;
 use Mvreisg\GamebaseBackend\Infrastructure\Connections\MariaDBConnection;
 use Mvreisg\GamebaseBackend\Infrastructure\Connections\RedisConnection;
 use Mvreisg\GamebaseBackend\Infrastructure\Encryption\SodiumEncryption;
 use Mvreisg\GamebaseBackend\Presentation\Controllers\GameGenreController;
-use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDBGameGenreRepository;
-use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDBUserRepository;
+use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDB\MariaDBGameGenreRepository;
+use Mvreisg\GamebaseBackend\Infrastructure\Repositories\MariaDB\MariaDBUserRepository;
 
 /**
  * Game Genre controller factory class.
