+++ tests/bootstrap.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Mvreisg\GamebaseBackend;
+
+use Dotenv\Dotenv;
+
+include_once dirname(__DIR__) . '/vendor/autoload.php';
+
+Dotenv::createImmutable(dirname(__DIR__))->load();
